requirements.txt
lxml
bs4
click
request
